Fixed 'escalating to SIGTERM' when closing
The issue was comming from the signal handler for the SIGINT event.
Indeed, it was always activated, and thus would catch our ctrl+C presses while not exiting the application.
Made it enabled only for the few lines where it is needed, and re-enabled the default one just after to make the application exit correctly